Commit a change after reorg and add efcore for the maui data project
