+++ src/BookShelves.Maui.Data/BookShelves.Maui.Data.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.8" />
     <PackageReference Include="sqlite-net-pcl" Version="1.9.172" />
 <!--    <PackageReference Include="SQLitePCLRaw.bundle_green" Version="2.1.9" />
 	<PackageReference Include="SQLitePCLRaw.lib.e_sqlite3" Version="2.1.2" PrivateAssets="None" GeneratePathProperty="true" Condition="'$(TargetFramework)'=='net8.0-maccatalyst'" />
