--- src/BookShelves.Maui.Data/SqliteProvider.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BookShelves.Maui.Data
-{
-    public static class SqliteProvider
-    {
-        public static void RegisterSqliteProvider()
-        {
-            SQLitePCL.raw.SetProvider(new SQLitePCL.SQLite3Provider_e_sqlite3());
-        }
-    }
-}
