+++ src/BookShelves.Maui/Services/SettingsService.cs
@@ -3,7 +3,7 @@
 
 using Microsoft.Extensions.Configuration;
 using Microsoft.Identity.Client.Extensions.Msal;
-using BookShelves.Shared.DataInterfaces;
+using BookShelves.Shared.ServiceInterfaces;
 
 namespace BookShelves.Maui.Services;
