+++ src/BookShelves.Shared.DataInterfaces/IItem.cs
