+++ src/BookShelves.Shared/Components/Layout/LayoutFooter.razor
@@ -1,4 +1,6 @@
-﻿@inject IVersionService VersionService
+﻿@using BookShelves.Shared.ServiceInterfaces
+@using BookShelves.Shared.ServiceModels
+@inject IVersionService VersionService
 
 <div class="footer nav-item px-3"> <!--navbar navbar-expand-lg">-->
     <div class="container-fluid">
