--- src/BookShelves.Shared/DataInterfaces/IVersionService.cs
@@ -1,7 +1,0 @@
-﻿namespace BookShelves.Shared.DataInterfaces;
-
-public interface IVersionService
-{
-    VersionInfo GetVersion();
-
-}
