+++ src/BookShelves.Shared/ServiceInterfaces/IVersionService.cs
@@ -1,0 +1,9 @@
+﻿using BookShelves.Shared.ServiceModels;
+
+namespace BookShelves.Shared.ServiceInterfaces;
+
+public interface IVersionService
+{
+    VersionInfo GetVersion();
+
+}
