+++ src/BookShelves.Shared/_Imports.razor
@@ -13,3 +13,4 @@
 @using BookShelves.Shared
 @using BookShelves.Shared.Components
 @using BookShelves.Shared.DataInterfaces
+@using BookShelves.Shared.ServiceInterfaces
